requirements.txt
h5py==2.9.0
matplotlib==2.2.4
numpy==1.16.4
Pillow==6.0.0
scikit-learn==0.20.3
torch==0.3.1
torchvision==0.2.0
